refactor: share filters with searchParams
