+++ frontend/packages/app/src/lib/constant.ts
@@ -3,6 +3,7 @@ export const ROUTES = {
   home: "/",
   task: "/task",
   project: "/projects",
+  "project-kanban": "/projects/kanban",
   "timesheet-personal": "/timesheet",
   "timesheet-team": "/timesheet/team",
   "timesheet-project": "/timesheet/project",
