+++ frontend/packages/app/src/pages/projects/kanban/provider.tsx
@@ -9,7 +9,8 @@ import { useFrappeGetCall, useFrappeUpdateDoc } from "frappe-react-sdk";
  */
 import { kebabToTitleCase } from "@/lib/utils";
 
-import { buildListFrappeFilters, useProjectFilter } from "../context";
+import { buildListFrappeFilters } from "../context";
+import { useProjectFilters } from "../hooks/useProjectFilters";
 import type { Phase } from "../types";
 import {
   ProjectKanbanContext,
@@ -18,7 +19,7 @@ import {
 import type { ResponseProjectKanban } from "./types";
 
 export function ProjectKanbanProvider({ children }: PropsWithChildren) {
-  const filters = useProjectFilter((c) => c.state.filters);
+  const { filters } = useProjectFilters();
 
   const frappeFilters = useMemo(
     () => buildListFrappeFilters(filters),
